# backend/calendar-api-swagger.yaml
openapi: 3.0.3
info:
  title: Calendar API
  description: |-
    This is an API for a calendar application. It allows you to manage your events and calendars with user authentication and database connection.

    Refer to this Github page for more information: [https://github.com/ssuntoso/calendar-web-app](https://github.com/ssuntoso/calendar-web-app)

    This API is built using Node.js and Express. It is also documented using Swagger UI.
  contact:
    email: animocainterview@gmail.com
  version: v1.0.8
servers:
  - url: https://calendar-api-114734088432.asia-east2.run.app/api/v1
...
paths:
...
  /event/addSubject:
    post:
      tags:
        - event
      summary: Add a new subject
      description: Add a new subject to the database
      operationId: addSubject
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                $ref: '#/components/schemas/Subject'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/MinimalSubject'
        '400':
          description: Request body must be an array of subjects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User does not have permission to add this subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error adding subject to database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /event/updateSubject:
    put:
      tags:
        - event
      summary: Update a subject
      description: Update a subject in the database
      operationId: updateSubject
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Subject'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/MinimalSubject'
        '403':
          description: User does not have permission to update this subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subject or User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error updating subject in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /event/deleteSubject:
    delete:
      tags:
        - event
      summary: Delete a subject
      description: Delete a subject in the database
      operationId: deleteSubject
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/DeleteSubject'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MinimalSubject'
        '403':
          description: User does not have permission to delete this subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subject or User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error deleting subject in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    Subject:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          example: 10
        subject_id:
          type: integer
          format: int64
          example: 1
        subject:
          type: string
          example: 'Meeting'
        start_time_zone:
          type: string
          example: 'Asia/Hong_Kong'
        start_time:
          type: string
          format: date-time
          example: '2021-08-01 10:00:00'
        end_time_zone:
          type: string
          example: 'Asia/Hong_Kong'
        end_time:
          type: string
          format: date-time
          example: '2021-08-01 11:00:00'
        all_day_event:
          type: boolean
          example: false
        description:
          type: string
          example: 'Meeting with John'
        location:
          type: string
          example: 'Office'
      xml:
        name: subject
    MinimalSubject:
      type: object
      properties:
        subject_id:
          type: integer
          format: int64
          example: 1
        subject:
          type: string
          example: 'Meeting'
      xml:
        name: minimalSubject
    DeleteSubject:
      type: object
      properties:
        subject_id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 10
      xml:
        name: deleteSubject
    Error:
      type: object
      properties:
        message:
          type: string
      xml:
        name: error
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: Enter the token return from login API with the `Bearer ` prefix, e.g. "Bearer abcde12345".